+++ src/main/java/com/ludo/study/studymatchingplatform/auth/naver/service/vo/property/OAuthProviderMapper.java
@@ -1,4 +1,4 @@
-package com.ludo.study.studymatchingplatform.auth.naver.service.naver.vo;
+package com.ludo.study.studymatchingplatform.auth.naver.service.vo.property;
 
 import java.util.HashMap;
 import java.util.Map;
